Changed plugin to use GlassFish 3.1-b40
Fix for adding all the specified dependencies to the server classpath. (issue http://java.net/jira/browse/GLASSFISH-15835)

Added fexibility to supply deploy/undeploy parameters. Currently, the plugin requires each deploy/undeploy parameter to be specified in its own element. With this change, any deployment paramter can be specified freely.

Eg., Before the fix:

<configuration>
<app>target/myapp.war</app>
<precompileJsp>true</precompileJsp>
<createTables>true</createTables>
..
</configuration>

After the fix:

<configuration>
<app>target/myapp.war</app>
<deploymentParams>
  <param>--precompilejsps=true</param>
  <param>--createtables=true</param>
</deploymentParams>
..
</configuration>

Same goes with undeployment.

Basically the idea is that any parameters that can be passed to 'asadmin deploy' can be passed as-it-is under <deploymentParams>

This change is backward compatible.



svn path=/branches/maven-plugin-3.1/; revision=44931